# docs/swagger.yaml
basePath: /
definitions:
...
  domain.Transaction:
    properties:
      account_id:
        type: string
      amount:
        type: integer
      bank_fee:
        type: integer
      category_id:
        type: string
      comment:
        type: string
      completed_at:
        type: string
      currency:
        type: string
      external_transaction_id:
        type: string
      is_hidden:
        type: boolean
      is_imported:
        type: boolean
      is_income:
        type: boolean
      mcc_code:
        type: string
      name_transaction:
        type: string
      receiver_account:
        type: string
      sender_account:
        type: string
      status:
        type: string
      transaction_id:
        type: string
      user_id:
        type: string
    type: object
...
  handler.CreateTransReq:
    properties:
      account_id:
        type: string
      amount:
        type: integer
      bank_fee:
        type: integer
      category_id:
        type: string
      comment:
        type: string
      completed_at:
        type: string
      currency:
        type: string
      is_income:
        type: boolean
      name:
        type: string
      status:
        type: string
    type: object
...
  handler.ToggleVisibilityReq:
    properties:
      hide:
        type: boolean
      transaction_ids:
        items:
          type: string
        type: array
    type: object
...
  handler.UpdateTransReq:
    properties:
      amount:
        type: integer
      bank_fee:
        type: integer
      category_id:
        type: string
      comment:
        type: string
      completed_at:
        type: string
      currency:
        type: string
      is_income:
        type: boolean
      name:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для приложения по управлению личными финансами.
  title: Finance Manager API
  version: "1.0"
paths:
...
  /api/v1/transactions:
    get:
      parameters:
      - description: ID счета
        in: query
        name: account_id
        type: string
      - description: ID категории
        in: query
        name: category_id
        type: string
      - description: Тип (доход/расход)
        in: query
        name: is_income
        type: boolean
      - description: Начальная дата (RFC3339)
        in: query
        name: start_date
        type: string
      - description: Конечная дата (RFC3339)
        in: query
        name: end_date
        type: string
      - description: Показать скрытые
        in: query
        name: is_hidden
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Transaction'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Получить транзакции (с фильтрами)
      tags:
      - transactions
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные транзакции
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateTransReq'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Создать транзакцию
      tags:
      - transactions
  /api/v1/transactions/{id}:
    delete:
      parameters:
      - description: ID транзакции
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить транзакцию
      tags:
      - transactions
    put:
      consumes:
      - application/json
      parameters:
      - description: ID транзакции
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateTransReq'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить транзакцию
      tags:
      - transactions
  /api/v1/transactions/visibility:
    patch:
      consumes:
      - application/json
      parameters:
      - description: IDs транзакций и статус
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ToggleVisibilityReq'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Изменить видимость транзакций
      tags:
      - transactions
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"